Note on list keys
React docs suggest that keys shouldn't be generated, but rather be embedded with data. Considering that list contents, usually, come from a database, this isn't a problem, in the vast majority of cases.

To comply with recommendations, instead of using a simple list of items, I'm using a list of objects, each one contains an id and a text.

Signed-off-by: Christian Molinari <chmolinari@christianmolinari.com>